requirements.txt
pydub~=0.25.1
azure-cognitiveservices-speech
